requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
binance==0.3
binance-connector==3.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dateparser==1.1.8
Django==4.2.2
django-scheduler==0.10.1
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
icalendar==5.0.7
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.2
plotly==5.15.0
pycparser==2.21
pycryptodome==3.18.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
spot==0.1a0
SQLAlchemy==2.0.16
sqlparse==0.4.4
tenacity==8.2.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
